#12008 [fix] 지표별SLA현황 목록 오류 수정

+++ src/main/kotlin/co/brainz/itsm/sla/metricYear/dto/MetricAnnualDto.kt
@@ -1,3 +1,9 @@
+/*
+ * Copyright 2020 Brainzcompany Co., Ltd.
+ * https://www.brainz.co.kr
+ *
+ */
+
 package co.brainz.itsm.sla.metricYear.dto
 
 import java.io.Serializable
@@ -8,7 +14,7 @@ data class MetricAnnualDto(
     val metricName: String? = "",
     val minValue: Double? = 0.0,
     val maxValue: Double? = 0.0,
-    val weightValue: Int = 0,
+    val weightValue: Double = 0.0,
     var score: Double? = 0.0,
     val owner: String? = "",
     val comment: String? = ""
